+++ CHANGELOG.md
@@ -13,6 +13,9 @@ Given a version number MAJOR.MINOR.PATCH, increment:
 
 
 ## [Unreleased]
+### Fixed
+- language attribute
+- parse functions
 
 ## [2.34.0] - 2025-07-11
 ### Added